# Always rebuild optTarget even when optimization driver is reused