Close #1888
optTarget used to be reset by resetComponent.
When resetComponent was remove and we don't reuse drivers, this was not a problem.
In #1830, we start to reuse driver for WF optimization and then caused a problem.
Now we ensure always rebuild optTarget even when optimization driver is reused.
The failures are only due to orbital rotation not ported to SoA.

```
ctest -j16 -R opt -E long
Test project /sandbox/opt/cleanup/yeluo/build_gnu_real
      Start 288: short-H4-opt-cslinear-linemin-16-1
 1/47 Test #288: short-H4-opt-cslinear-linemin-16-1 .................   Passed   56.74 sec
      Start 274: short-H2-orb-opt-16-1
 2/47 Test #274: short-H2-orb-opt-16-1 ..............................***Failed    0.22 sec
      Start 278: short-H4-opt-linear-16-1
 3/47 Test #278: short-H4-opt-linear-16-1 ...........................   Passed   51.80 sec
      Start 280: short-H4-opt-linear-rescale-4-4
 4/47 Test #280: short-H4-opt-linear-rescale-4-4 ....................   Passed   31.26 sec
      Start 282: short-H4-opt-linear-linemin-16-1
 5/47 Test #282: short-H4-opt-linear-linemin-16-1 ...................   Passed   40.56 sec
      Start 284: short-H4-opt-cslinear-quartic-4-4
 6/47 Test #284: short-H4-opt-cslinear-quartic-4-4 ..................   Passed   55.60 sec
      Start 286: short-H4-opt-cslinear-rescale-16-1
 7/47 Test #286: short-H4-opt-cslinear-rescale-16-1 .................   Passed   36.88 sec
      Start 290: short-H4-opt-adaptive-16-1
 8/47 Test #290: short-H4-opt-adaptive-16-1 .........................   Passed   52.84 sec
      Start 292: short-H4-opt-adaptive-1-16
 9/47 Test #292: short-H4-opt-adaptive-1-16 .........................   Passed   63.77 sec
      Start 294: short-H4-orb-opt-16-1
10/47 Test #294: short-H4-orb-opt-16-1 ..............................***Failed    0.10 sec
      Start 296: short-H4-orb-opt-dmc-16-1
11/47 Test #296: short-H4-orb-opt-dmc-16-1 ..........................***Failed    0.21 sec
      Start 298: short-H4-orb-opt-dmc-1-16
12/47 Test #298: short-H4-orb-opt-dmc-1-16 ..........................***Failed    0.27 sec
      Start 304: short-H4-opt-OneShiftOnly-16-1
13/47 Test #304: short-H4-opt-OneShiftOnly-16-1 .....................   Passed   26.08 sec
      Start 306: short-H4-opt-OneShiftOnly-1-16
14/47 Test #306: short-H4-opt-OneShiftOnly-1-16 .....................   Passed   35.14 sec
      Start 308: short-H4-msd-orb-opt-16-1
15/47 Test #308: short-H4-msd-orb-opt-16-1 ..........................   Passed   16.50 sec
      Start 389: short-LiH_ae-vmc_msdj_optCI-1-16
16/47 Test #389: short-LiH_ae-vmc_msdj_optCI-1-16 ...................   Passed   64.29 sec
      Start 583: short-diamondC_1x1x1_pp-opt_sdj-1-16
17/47 Test #583: short-diamondC_1x1x1_pp-opt_sdj-1-16 ...............   Passed   73.29 sec
      Start 1045: example_He_simple_opt-1-1
      Start  22: deterministic-unit_test_optimize
      Start 289: short-H4-opt-cslinear-linemin-16-1-totenergy
      Start 275: short-H2-orb-opt-16-1-totenergy
      Start 279: short-H4-opt-linear-16-1-totenergy
      Start 281: short-H4-opt-linear-rescale-4-4-totenergy
      Start 283: short-H4-opt-linear-linemin-16-1-totenergy
      Start 285: short-H4-opt-cslinear-quartic-4-4-totenergy
      Start 287: short-H4-opt-cslinear-rescale-16-1-totenergy
      Start 291: short-H4-opt-adaptive-16-1-totenergy
      Start 293: short-H4-opt-adaptive-1-16-totenergy
      Start 295: short-H4-orb-opt-16-1-totenergy
      Start 297: short-H4-orb-opt-dmc-16-1-totenergy
      Start 299: short-H4-orb-opt-dmc-1-16-totenergy
      Start 305: short-H4-opt-OneShiftOnly-16-1-totenergy
      Start 307: short-H4-opt-OneShiftOnly-1-16-totenergy
18/47 Test #275: short-H2-orb-opt-16-1-totenergy ....................***Failed    0.03 sec
      Start 309: short-H4-msd-orb-opt-16-1-totenergy
19/47 Test #283: short-H4-opt-linear-linemin-16-1-totenergy .........   Passed    0.03 sec
      Start 390: short-LiH_ae-vmc_msdj_optCI-1-16-kinetic
20/47 Test #279: short-H4-opt-linear-16-1-totenergy .................   Passed    0.04 sec
      Start 391: short-LiH_ae-vmc_msdj_optCI-1-16-totenergy
21/47 Test #285: short-H4-opt-cslinear-quartic-4-4-totenergy ........   Passed    0.04 sec
      Start 392: short-LiH_ae-vmc_msdj_optCI-1-16-variance
22/47 Test #287: short-H4-opt-cslinear-rescale-16-1-totenergy .......   Passed    0.04 sec
      Start 393: short-LiH_ae-vmc_msdj_optCI-1-16-eeenergy
23/47 Test #293: short-H4-opt-adaptive-1-16-totenergy ...............   Passed    0.03 sec
      Start 394: short-LiH_ae-vmc_msdj_optCI-1-16-potential
24/47 Test #281: short-H4-opt-linear-rescale-4-4-totenergy ..........   Passed    0.05 sec
      Start 395: short-LiH_ae-vmc_msdj_optCI-1-16-ionion
25/47 Test #299: short-H4-orb-opt-dmc-1-16-totenergy ................***Failed    0.03 sec
      Start 584: short-diamondC_1x1x1_pp-opt_sdj-1-16-kinetic
26/47 Test #295: short-H4-orb-opt-16-1-totenergy ....................***Failed    0.03 sec
      Start 585: short-diamondC_1x1x1_pp-opt_sdj-1-16-totenergy
27/47 Test #289: short-H4-opt-cslinear-linemin-16-1-totenergy .......   Passed    0.05 sec
      Start 586: short-diamondC_1x1x1_pp-opt_sdj-1-16-samples
28/47 Test #297: short-H4-orb-opt-dmc-16-1-totenergy ................***Failed    0.03 sec
      Start 587: short-diamondC_1x1x1_pp-opt_sdj-1-16-localecp
29/47 Test #307: short-H4-opt-OneShiftOnly-1-16-totenergy ...........   Passed    0.03 sec
      Start 588: short-diamondC_1x1x1_pp-opt_sdj-1-16-nonlocalecp
30/47 Test #305: short-H4-opt-OneShiftOnly-16-1-totenergy ...........   Passed    0.04 sec
      Start 1053: ntest_nexus_optional_dependencies
31/47 Test #291: short-H4-opt-adaptive-16-1-totenergy ...............   Passed    0.05 sec
32/47 Test  #22: deterministic-unit_test_optimize ...................   Passed    0.06 sec
33/47 Test #309: short-H4-msd-orb-opt-16-1-totenergy ................   Passed    0.02 sec
34/47 Test #390: short-LiH_ae-vmc_msdj_optCI-1-16-kinetic ...........   Passed    0.03 sec
35/47 Test #391: short-LiH_ae-vmc_msdj_optCI-1-16-totenergy .........   Passed    0.02 sec
36/47 Test #392: short-LiH_ae-vmc_msdj_optCI-1-16-variance ..........   Passed    0.03 sec
37/47 Test #393: short-LiH_ae-vmc_msdj_optCI-1-16-eeenergy ..........   Passed    0.03 sec
38/47 Test #584: short-diamondC_1x1x1_pp-opt_sdj-1-16-kinetic .......   Passed    0.02 sec
39/47 Test #585: short-diamondC_1x1x1_pp-opt_sdj-1-16-totenergy .....   Passed    0.02 sec
40/47 Test #586: short-diamondC_1x1x1_pp-opt_sdj-1-16-samples .......   Passed    0.02 sec
41/47 Test #395: short-LiH_ae-vmc_msdj_optCI-1-16-ionion ............   Passed    0.04 sec
42/47 Test #588: short-diamondC_1x1x1_pp-opt_sdj-1-16-nonlocalecp ...   Passed    0.03 sec
43/47 Test #394: short-LiH_ae-vmc_msdj_optCI-1-16-potential .........   Passed    0.04 sec
44/47 Test #587: short-diamondC_1x1x1_pp-opt_sdj-1-16-localecp ......   Passed    0.03 sec
45/47 Test #1053: ntest_nexus_optional_dependencies ..................   Passed    0.10 sec
46/47 Test #1045: example_He_simple_opt-1-1 ..........................   Passed   29.30 sec
      Start 1046: example_He_simple_opt-1-1-totenergy
47/47 Test #1046: example_He_simple_opt-1-1-totenergy ................   Passed    0.04 sec

83% tests passed, 8 tests failed out of 47

Label Time Summary:
QMCPACK                     = 9718.46 sec*proc (18 tests)
QMCPACK-checking-results    =   0.89 sec*proc (27 tests)
abort                       =  13.01 sec*proc (8 tests)
bug                         =  13.01 sec*proc (8 tests)
cause_unknown               = 2456.13 sec*proc (6 tests)
deterministic               =   0.06 sec*proc (1 test)
intermittent_stat_fail      = 2456.13 sec*proc (6 tests)
nexus                       =   0.10 sec*proc (1 test)
quality_unknown             = 9719.52 sec*proc (47 tests)
unit                        =   0.06 sec*proc (1 test)
unstable                    =   0.25 sec*proc (7 tests)

Total Test time (real) = 635.17 sec

The following tests FAILED:
	274 - short-H2-orb-opt-16-1 (Failed)
	275 - short-H2-orb-opt-16-1-totenergy (Failed)
	294 - short-H4-orb-opt-16-1 (Failed)
	295 - short-H4-orb-opt-16-1-totenergy (Failed)
	296 - short-H4-orb-opt-dmc-16-1 (Failed)
	297 - short-H4-orb-opt-dmc-16-1-totenergy (Failed)
	298 - short-H4-orb-opt-dmc-1-16 (Failed)
	299 - short-H4-orb-opt-dmc-1-16-totenergy (Failed)
Errors while running CTest
```